move common steps to base commonSteps.js
